SAS-33: #changed addBudget in class Season, and added imports to classes Team and Game

+++ src/main/java/com/SAS/League/Budget.java
@@ -1,5 +1,7 @@
 package com.SAS.League;
 
+import com.SAS.team.Team;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
@@ -7,7 +9,7 @@ import java.util.Map;
 public class Budget {
     private Team team;
     private Season season;
-    private int budget;
+    private double budget;
 
     /**
      * GamesArrangment constructor with attributes
@@ -16,7 +18,7 @@ public class Budget {
      * @param season: the budget's season
      * @param budget: the list of the games
      */
-    public Budget(Team team, Season season, int budget) {
+    public Budget(Team team, Season season, double budget) {
         this.team = team;
         this.season = season;
         this.budget = budget;
@@ -42,7 +44,7 @@ public class Budget {
     /**
      * @return the size of the budget
      */
-    public int getBudget() {
+    public double getBudget() {
         return budget;
     }
